+++ src/mvvm/awtk/navigator_handler_awtk.c
@@ -31,12 +31,15 @@ static const object_vtable_t s_navigator_handler_awtk_vtable = {
 
 static ret_t navigator_handler_awtk_on_request(navigator_handler_t* handler,
                                                navigator_request_t* req) {
+  model_t* model = NULL;
   navigator_handler_awtk_t* handler_awtk = NAVIGATOR_HANDLER_AWTK(handler);
-  model_t* model = handler_awtk->model_create(handler_awtk->model_create_args);
-  return_value_if_fail(model != NULL, RET_FAIL);
 
-  ENSURE(model_on_will_mount(model, req) == RET_OK);
-  return vm_open_window(req->target, model);
+  if (handler_awtk->model_create != NULL) {
+    model = handler_awtk->model_create(handler_awtk->model_create_args);
+    return_value_if_fail(model != NULL, RET_FAIL);
+  }
+
+  return vm_open_window(req->target, model, req);
 }
 
 navigator_handler_t* navigator_handler_awtk_create(model_create_t model_create,
@@ -53,6 +56,21 @@ navigator_handler_t* navigator_handler_awtk_create(model_create_t model_create,
 
   handler_awtk->model_create = model_create;
   handler_awtk->model_create_args = model_create_args;
+  handler->on_request = navigator_handler_awtk_on_request;
+
+  return handler;
+}
+
+navigator_handler_t* navigator_handler_awtk_default_create(void) {
+  object_t* obj = NULL;
+  navigator_handler_t* handler = NULL;
+  navigator_handler_awtk_t* handler_awtk = NULL;
+
+  obj = object_create(&s_navigator_handler_awtk_vtable);
+  handler = NAVIGATOR_HANDLER(obj);
+  handler_awtk = NAVIGATOR_HANDLER_AWTK(obj);
+  return_value_if_fail(handler != NULL, NULL);
+
   handler->on_request = navigator_handler_awtk_on_request;
 
   return handler;