+++ src/mvvm/base/model_factory.c
@@ -1,0 +1,96 @@
+﻿/**
+ * File:   model_factory.c
+ * Author: AWTK Develop Team
+ * Brief:  model factory
+ *
+ * Copyright (c) 2019 - 2019  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2019-02-08 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#include "tkc/mem.h"
+#include "tkc/utils.h"
+#include "tkc/object_default.h"
+#include "mvvm/base/model_factory.h"
+
+static model_factory_t* s_model_factory;
+
+ret_t model_factory_init(void) {
+  if (s_model_factory == NULL) {
+    s_model_factory = TKMEM_ZALLOC(model_factory_t);
+    return_value_if_fail(s_model_factory != NULL, RET_OOM);
+
+    s_model_factory->creators = object_default_create();
+    if (s_model_factory->creators == NULL) {
+      TKMEM_FREE(s_model_factory);
+      s_model_factory = NULL;
+    }
+  }
+
+  return s_model_factory != NULL ? RET_OK : RET_FAIL;
+}
+
+bool_t model_factory_exist(const char* type) {
+  return_value_if_fail(s_model_factory != NULL && type != NULL, FALSE);
+
+  return object_has_prop((s_model_factory->creators), type);
+}
+
+ret_t model_factory_unregister(const char* type) {
+  return_value_if_fail(s_model_factory != NULL && type != NULL, RET_BAD_PARAMS);
+
+  return object_remove_prop((s_model_factory->creators), type);
+}
+
+ret_t model_factory_register(const char* type, model_create_t create) {
+  return_value_if_fail(s_model_factory != NULL && type != NULL && create != NULL, RET_BAD_PARAMS);
+
+  return object_set_prop_pointer(s_model_factory->creators, type, create);
+}
+
+model_t* model_factory_create(const char* type, void* args) {
+  model_create_t create = NULL;
+  return_value_if_fail(s_model_factory != NULL && type != NULL && args != NULL, NULL);
+  create = (model_create_t)object_get_prop_pointer(s_model_factory->creators, type);
+  return_value_if_fail(create != NULL, NULL);
+
+  return create(args);
+}
+
+model_t* model_factory_create_with_filename(const char* filename) {
+  char name[MAX_PATH + 1];
+  char* ext_name = NULL;
+  return_value_if_fail(filename != NULL, NULL);
+  tk_strncpy(name, filename, MAX_PATH);
+
+  ext_name = strrchr(name, '.');
+  return_value_if_fail(ext_name != NULL, NULL);
+
+  *ext_name = '\0';
+  ext_name++;
+
+  return model_factory_create(ext_name, name);
+}
+
+ret_t model_factory_deinit(void) {
+  return_value_if_fail(s_model_factory != NULL && s_model_factory->creators != NULL,
+                       RET_BAD_PARAMS);
+
+  object_unref(s_model_factory->creators);
+  TKMEM_FREE(s_model_factory);
+
+  s_model_factory = NULL;
+
+  return RET_OK;
+}